refactor(profile-review): extract legacy page and add feature flag routing
