+++ app/admin/profile-review/profile-review-v2.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+export default function ProfileReviewV2() {
+  return <div>Profile Review V2 (placeholder)</div>;
+}
